backend/pyproject.toml
[project]
name = "datak-gateway"
version = "0.1.0"
description = "Industrial IoT Edge Gateway & Data Aggregator"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "AGPL-3.0"}
authors = [
    {name = "DaTaK Team", email = "dev@datak.io"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "influxdb-client[async]>=1.39.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "RestrictedPython>=7.0",
    
    # Industrial Protocols
    "pymodbus>=3.6.0",
    "aiomqtt>=2.0.0",
    "python-can>=4.3.0",
    "cantools>=39.4.0",
    
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    
    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    
    # Utilities
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
datak = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
